+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to occ-tools will be documented in this file.
 This project must adhere to [this](https://github.com/olivierlacan/keep-a-changelog/blob/master/CHANGELOG.md) format.
 
+## [2.0.0-beta.42] - 2021-05-12]
+### Added
+- Option to append a "banner" in the app-level, in that way, we can define some global JS and configure requirejs aliases properly.
+- "project-settings" in the occ-tools.project.json. This will file can be used as a default setting for the project.
+- Added "ASSETS_VERSION" global config, this value can be used to set the files version. Webpack will replace all occurencies of ASSETS_VERSION in the code.
 ## [2.0.0-beta.41] - 2021-05-01]
 ### Added
 - Minify `.json` and `.js` extensions when uploading files